django_backend/pyproject.toml
[project]
name = "manac-backend"
version = "1.0.0"
description = "ManAC - Campus Equipment Management Backend"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14,<4.0",
    "django-cors-headers>=4.3,<5.0",
    "gunicorn>=21.0,<22.0",
    "psycopg2-binary>=2.9,<3.0",
    "python-dotenv>=1.0,<2.0",
    "Pillow>=10.0,<11.0",
    "dj-database-url>=2.0,<3.0",
    "django-filter>=2.4,<3.0",
    "dj-database-url"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["manac", "accounts", "departments", "equipment", "borrowings"]

[tool.vercel]
django = "manac.wsgi.application"
